+++ src/data/types.ts
@@ -21,5 +21,6 @@ export type TokenID = Tables['tokens']['Row']['id'];
 export type TransactionStatus = Database['public']['Enums']['transaction_status'];
 export type SubmissionRow = Tables['submissions']['Row'];
 export type SubmissionID = SubmissionRow['id'];
+export type SubmissionStatus = Database['public']['Enums']['submission_status'];
 export type AssignmentRow = Tables['assignments']['Row'];
 export type AssignmentID = AssignmentRow['id'];